Mcp_Chatbot/backend/requirements-vector.txt
# LangChain Core
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-pinecone>=0.0.3
langchain-community>=0.0.10

# Vector Database & Embeddings
pinecone-client>=3.0.0
openai>=1.0.0

# Document Processing
unstructured>=0.10.0
unstructured[md]>=0.10.0

# API & Backend
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Additional utilities
tqdm>=4.65.0
python-multipart>=0.0.6
tiktoken>=0.5.0
